+++ mrmustard/math/lattice/strategies/compactFock/diagonal_amps.py
@@ -15,7 +15,9 @@ from mrmustard.math.lattice.strategies.compactFock.helperFunctions import (
 
 
 @njit
-def use_offDiag_pivot(A, B, M, cutoffs, params, d, arr0, arr2, arr1010, arr1001, arr1):
+def use_offDiag_pivot(
+    A, B, M, cutoffs, params, d, arr0, arr2, arr1010, arr1001, arr1
+):  # pragma: no cover
     """
     Apply recurrence relation for pivot of type [a+1,a,b,b,c,c,...] / [a,a,b+1,b,c,c,...] / [a,a,b,b,c+1,c,...]
     Args:
@@ -33,7 +35,10 @@ def use_offDiag_pivot(A, B, M, cutoffs, params, d, arr0, arr2, arr1010, arr1001,
     pivot[2 * d] += 1
     K_l = SQRT[pivot]
     K_i = SQRT[pivot + 1]
-    G_in = np.empty(2 * M, dtype=np.complex128)
+    if B.ndim == 1:
+        G_in = np.empty(2 * M, dtype=np.complex128)
+    elif B.ndim == 2:
+        G_in = np.empty((2 * M, B.shape[1]), dtype=np.complex128)
 
     ########## READ ##########
     GB = arr1[(2 * d,) + params] * B
@@ -54,7 +59,10 @@ def use_offDiag_pivot(A, B, M, cutoffs, params, d, arr0, arr2, arr1010, arr1001,
             G_in[2 * i + 1] = arr1010[(d, i - d - 1) + params_adapted]
 
     ########## WRITE ##########
-    G_in = np.multiply(K_l, G_in)
+    if B.ndim == 1:
+        G_in = np.multiply(K_l, G_in)
+    elif B.ndim == 2:
+        G_in = np.multiply(np.expand_dims(K_l, 1), G_in)
 
     # Array0
     params_adapted = tuple_setitem(params, d, params[d] + 1)
@@ -76,7 +84,7 @@ def use_offDiag_pivot(A, B, M, cutoffs, params, d, arr0, arr2, arr1010, arr1001,
 
 
 @njit
-def use_diag_pivot(A, B, M, cutoffs, params, arr0, arr1):
+def use_diag_pivot(A, B, M, cutoffs, params, arr0, arr1):  # pragma: no cover
     """
     Apply recurrence relation for pivot of type [a,a,b,b,c,c...]
     Args:
@@ -91,7 +99,10 @@ def use_diag_pivot(A, B, M, cutoffs, params, arr0, arr1):
     pivot = repeat_twice(params)
     K_l = SQRT[pivot]
     K_i = SQRT[pivot + 1]
-    G_in = np.empty(2 * M, dtype=np.complex128)
+    if B.ndim == 1:
+        G_in = np.empty(2 * M, dtype=np.complex128)
+    elif B.ndim == 2:
+        G_in = np.empty((2 * M, B.shape[1]), dtype=np.complex128)
 
     ########## READ ##########
     GB = arr0[params] * B
@@ -105,7 +116,10 @@ def use_diag_pivot(A, B, M, cutoffs, params, arr0, arr1):
             ]  # [i+1-2*(i%2) for i in range(6)] = [1,0,3,2,5,4]
 
     ########## WRITE ##########
-    G_in = np.multiply(K_l, G_in)
+    if B.ndim == 1:
+        G_in = np.multiply(K_l, G_in)
+    elif B.ndim == 2:
+        G_in = np.multiply(np.expand_dims(K_l, 1), G_in)
 
     # Array1
     for i in range(2 * M):
@@ -120,7 +134,7 @@ def use_diag_pivot(A, B, M, cutoffs, params, arr0, arr1):
 @njit
 def fock_representation_diagonal_amps_NUMBA(
     A, B, M, cutoffs, arr0, arr2, arr1010, arr1001, arr1, tuple_type, list_type
-):
+):  # pragma: no cover
     """
     Returns the PNR probabilities of a mixed state according to algorithm 1 of:
     https://doi.org/10.22331/q-2023-08-29-1097
@@ -165,16 +179,30 @@ def fock_representation_diagonal_amps(A, B, G0, M, cutoffs):
     tuple_type = numba.types.UniTuple(int64, M)
     list_type = numba.types.ListType(tuple_type)
 
-    arr0 = np.empty(cutoffs, dtype=np.complex128)
+    if B.ndim == 1:
+        arr0 = np.empty(cutoffs, dtype=np.complex128)
+        arr2 = np.empty((M,) + cutoffs, dtype=np.complex128)
+        arr1 = np.empty((2 * M,) + cutoffs, dtype=np.complex128)
+        if M == 1:
+            arr1010 = np.empty((1, 1, 1), dtype=np.complex128)
+            arr1001 = np.empty((1, 1, 1), dtype=np.complex128)
+        else:
+            arr1010 = np.empty((M, M - 1) + cutoffs, dtype=np.complex128)
+            arr1001 = np.empty((M, M - 1) + cutoffs, dtype=np.complex128)
+
+    elif B.ndim == 2:
+        batch_length = B.shape[1]
+        arr0 = np.empty(cutoffs + (batch_length,), dtype=np.complex128)
+        arr2 = np.empty((M,) + cutoffs + (batch_length,), dtype=np.complex128)
+        arr1 = np.empty((2 * M,) + cutoffs + (batch_length,), dtype=np.complex128)
+        if M == 1:
+            arr1010 = np.empty((1, 1, 1) + (batch_length,), dtype=np.complex128)
+            arr1001 = np.empty((1, 1, 1) + (batch_length,), dtype=np.complex128)
+        else:
+            arr1010 = np.empty((M, M - 1) + cutoffs + (batch_length,), dtype=np.complex128)
+            arr1001 = np.empty((M, M - 1) + cutoffs + (batch_length,), dtype=np.complex128)
+
     arr0[(0,) * M] = G0
-    arr2 = np.empty((M,) + cutoffs, dtype=np.complex128)
-    arr1 = np.empty((2 * M,) + cutoffs, dtype=np.complex128)
-    if M == 1:
-        arr1010 = np.empty((1, 1, 1), dtype=np.complex128)
-        arr1001 = np.empty((1, 1, 1), dtype=np.complex128)
-    else:
-        arr1010 = np.empty((M, M - 1) + cutoffs, dtype=np.complex128)
-        arr1001 = np.empty((M, M - 1) + cutoffs, dtype=np.complex128)
     return fock_representation_diagonal_amps_NUMBA(
         A, B, M, cutoffs, arr0, arr2, arr1010, arr1001, arr1, tuple_type, list_type
     )